requirements.txt
strands-agents
strands-agents-tools
ddgs
